+++ CHANGELOG.md
@@ -59,6 +59,7 @@ these, please run the `LFPBandV1().fix_1481()` method after updating.
 - Improve export process for speed and generalization #1387
 - Additional methods for updating files for DANDI standards #1387
 - Implementation of union and intersect methods for restriction graphs #1387
+- Add file issue checks to AnalysisNwbfile cleanup steps #1431
 
 ### Pipelines